Add files via upload
Please take a look at this screenshot. it shows for the portion under "Cluster Cryptocurrencies with K-means using the scaled PCA data frame" for the fit of the k-mean model using the PCA data it pops up. uploading it to GitHub gives an error that it can't read the HTML.